chore: Move token_counter.py
