+++ benchmarks/page-token-benchmark/src/token_counter.py
